fix(format): enforce little-endian float64, nanosecond timestamps, remove ROS workaround
- Remove ROS pytest workaround from pyproject.toml (testpaths only)
- Enforce little-endian float64 dtype in states_actions.bin (writer/reader)
- Camera timestamps now use nanosecond timeline (t0 + tick*index) instead of frame indices
- Strengthen roundtrip test: verify dtype uint8 and shape (T, 16, 16, 3) for decoded frames

All 8 tests pass without addopts workaround.

+++ abcdl/format/writer.py
@@ -19,7 +19,7 @@ def write_abcdl(episode: Episode, out_dir: str) -> None:
 
     sa = np.concatenate([np.asarray(episode.states, np.float64),
                          np.asarray(episode.actions, np.float64)], axis=1)
-    sa.tofile(os.path.join(out_dir, "states_actions.bin"))
+    sa.astype("<f8").tofile(os.path.join(out_dir, "states_actions.bin"))
 
     stacks = []
     h = w = None
